feat: Refactored project
-Added multiple new dependencies.
-Added pyproject.toml
-Moved the entry point to the app in
app/__main__.py
-Adding a tests folder
-The refactor for this project was done
 to help keep it clean and easy to maintain.
