+++ .gitignore
@@ -5,7 +5,8 @@ __pycache__/
 
 # C extensions
 *.so
-
+#uv
+uv.lock
 # Distribution / packaging
 .Python
 build/
