+++ src/config/finance_manager.json
@@ -1,0 +1,3 @@
+{
+    "monthly_salary": []
+}
