+++ src/processing/cleaners.py
@@ -1,0 +1,53 @@
+"""The cleaners module contains a class for cleaning and normalized csv data.
+"""
+from pathlib import Path
+from typing import List
+import pandas as pd
+from config import normalization_config
+
+
+class CSVDataCleaner:
+    """Cleans and normalizes incoming CSV finance data."""
+
+    def __init__(self, path: Path):
+        self.path = Path(path)
+        self._columns_map = normalization_config["columns_map"]
+        self._drop_list = normalization_config["drop_list"]
+
+    def load_and_clean(self) -> List[dict]:
+        """Load and clean all CSV files in a directory.
+
+        Returns:
+            list: {account_type: cleaned DataFrame}
+        """
+        if not self.path.is_dir():
+            raise ValueError(f"{self.path} is not a directory")
+
+        cleaned = []
+        for file in self.path.glob("*.csv"):
+            record = {}
+            df = pd.read_csv(file)
+            key = "credit" if "Transaction Date" in df.columns else "checking"
+            df = self._clean_dataframe(df)
+            record[key] = df
+            cleaned.append(record)
+        return cleaned
+
+    def _clean_dataframe(self, df: pd.DataFrame) -> pd.DataFrame:
+        """Apply cleaning rules to a single DataFrame."""
+        if "Type" in df.columns:
+            df = df[df["Type"] == "Sale"]
+        if "Date" in df.columns:
+            df["Date"] = pd.to_datetime(df["Date"], errors="coerce")
+        if "Amount" in df.columns:
+            df.loc[:, "Amount"] = pd.to_numeric(
+                df["Amount"], errors="coerce").fillna(0)
+        for col in self._drop_list:
+            if col in df.columns:
+                df = df.drop(columns=[col])
+        df = df.rename(columns=self._columns_map)
+        return df.iloc[:, 0:4]
+
+
+if __name__ == "__main__":
+    pass
